Restore standard Ctrl + Delete behavior and move clear note shortcut to Ctrl + Shift + Delete

+++ js/app.js
@@ -1456,7 +1456,7 @@ you can buy me a coffee — the link of which is available in the About section.
 			copyNotesToClipboard(notepad.note.val());
 		}
 
-		if ((event.ctrlKey && event.code === 'Delete') || (event.metaKey && event.code === 'Backspace')) {
+		if ((event.ctrlKey && event.shiftKey && event.code === 'Delete') || (event.metaKey && event.code === 'Backspace')) {
 			event.preventDefault();
 			deleteNotes();
 		}
